% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_p_xxx.R
\name{ci_p_add_4}
\alias{ci_p_add_4}
\title{Add-4 Wald-t Confidence Interval for Binomial Proportion}
\usage{
ci_p_add_4(x, n, conf.level = 0.95)
}
\arguments{
\item{x}{A number or a vector with the number of successes.}

\item{n}{A number or a vector with the number of trials.}

\item{conf.level}{Confidence level for the returned confidence interval. By default, it is 0.95.}
}
\value{
A vector with the lower and upper limits of the confidence interval.
}
\description{
This function calculates the Add-4 Wald-t confidence interval for a Binomial proportion using the method XX proposed by Pan (2002).
It modifies the classical Wald interval by adding four pseudo-observations and uses a t-distribution with adjusted degrees of freedom \eqn{\nu}, as specified in equation (2.9).
The function is vectorized, allowing for evaluation of single values or vectors.
}
\details{
The Add-4 Wald-t confidence interval improves the performance of the Wald interval by adding 2 successes and 2 failures to the observed data, effectively modifying the estimated proportion:

\deqn{\hat{p} = \frac{x + 2}{n + 4}.}

The variance \eqn{V(\hat{p}, n+4)} is given by:

\deqn{V(\hat{p}, n+4) = \frac{\hat{p}(1 - \hat{p})}{n + 4}.}

The degrees of freedom \eqn{\nu} are calculated using equation (2.9):

\deqn{\nu = \frac{2 V(\hat{p}, n+4)^2}{\Omega(\hat{p}, n+4)},}

where \eqn{\Omega(\hat{p}, n+4)} is defined as:

\deqn{\Omega(p, n) = \frac{p - p^2}{n^3} + \frac{p + (6n - 7)p^2 + 4(n - 1)(n - 3)p^3 - 2(n - 1)(2n - 3)p^4}{n^5} - \frac{2(p + (2n - 3)p^2 - 2(n - 1)p^3)}{n^4}.}

The confidence interval is then calculated as:

\deqn{\text{Lower} = \hat{p} - t \cdot \sqrt{\frac{\tilde{\pi}(1 - \hat{p})}{n+4}},}
\deqn{\text{Upper} = \hat{p} + t \cdot \sqrt{\frac{\tilde{\pi}(1 - \hat{p})}{n+4}},}

where \eqn{t} is the critical value from the t-distribution with \eqn{\nu} degrees of freedom.
}
\examples{
ci_p_add_4(x=15, n=50, conf.level=0.95)
ci_p_add_4(x=0,  n=50, conf.level=0.95)
ci_p_add_4(x=50, n=50, conf.level=0.95)

}
\references{
Missing reference.
}
\seealso{
\link{ci_p}.
}
\author{
David Esteban Cartagena Mejía, \email{dcartagena@unal.edu.co}
}
