% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_p_xxx.R
\name{ci_p_agresti_caffo}
\alias{ci_p_agresti_caffo}
\title{Agresti-Caffo confidence interval for binomial proportion}
\usage{
ci_p_agresti_caffo(x, n, conf.level = 0.95)
}
\arguments{
\item{x}{A number or a vector with the number of successes.}

\item{n}{A number or a vector with the number of trials.}

\item{conf.level}{Confidence level for the returned confidence interval.
By default, it is 0.95.}
}
\value{
A vector with the lower and upper limits of the confidence interval.
}
\description{
This function calculates the Agresti-Caffo confidence interval for
a Binomial
proportion, which includes a Bayesian adjustment by adding 2 successes
and 2 failures to stabilize the estimate. It is vectorized,
allowing the evaluation of single values or vectors.
}
\details{
The Agresti-Caffo confidence interval incorporates a simple
Bayesian adjustment by adding 2 successes and 2 failures
to the data. The adjusted proportion is calculated as:

\deqn{\hat{p} = \frac{x + 2}{n + 4}}

The standard error for the adjusted proportion is given by:

\deqn{\text{se} = \sqrt{\frac{\hat{p} (1 - \hat{p})}{n + 4}}.}

The confidence interval is then constructed using the \eqn{t}-distribution
with \eqn{n - 1} degrees of freedom:

\deqn{\text{CI} = \hat{p} \pm t_{n-1, \alpha/2} \cdot \text{se},}

where \eqn{t_{n-1, \alpha/2}} is the critical value of the
\eqn{t}-distribution at a two-tailed significance level of
\eqn{\alpha = 1 - \text{conf.level}}.
}
\examples{
# Example with a single value
ci_p_agresti_caffo(x = 15, n = 50, conf.level = 0.95)


}
\references{
Agresti, Alan, and Brian Caffo. "Simple and effective confidence
intervals for proportions and differences of proportions result
from adding two successes and two failures." The American
Statistician 54.4 (2000): 280-288.
}
\seealso{
\link{ci_p}.
}
\author{
David Esteban Cartagena Mejía, \email{dcartagena@unal.edu.co}
}
