% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_p_xxx.R
\name{ci_p_agresti_coull}
\alias{ci_p_agresti_coull}
\title{Agresti-Coull confidence interval for binomial proportion}
\usage{
ci_p_agresti_coull(x, n, conf.level = 0.95)
}
\arguments{
\item{x}{A number or a vector with the number of successes.}

\item{n}{A number or a vector with the number of trials.}

\item{conf.level}{Confidence level for the returned confidence interval.
By default, it is 0.95.}
}
\value{
A vector with the lower and upper limits.
}
\description{
This function calculates the confidence interval for a proportion.
It is vectorized, allowing users to evaluate it using either
single values or vectors.
}
\details{
The Agresti-Coull interval is an approximate confidence interval for the
binomial proportion \eqn{p}.
The limits are calculated based on an adjusted proportion \eqn{\tilde{p}}
and its standard error. The mathematical definitions are as follows:
 Adjusted proportion: \eqn{\tilde{p}=\frac{x + 2}{n + 4}};
 Adjusted standard error: \eqn{se=\sqrt{\frac{\tilde{p}(1 - \tilde{p})}{n + 4}}};
 Confidence limits: \eqn{\tilde{p} \pm z_{\alpha/2} \cdot se},

where \eqn{z_{\alpha/2}} is the critical value of the standard normal
distribution. The limits are truncated to the range \eqn{[0, 1]}.
}
\examples{
ci_p_agresti_coull(x=15, n=50, conf.level=0.95)

}
\references{
Agresti, A., & Coull, B. A. (1998). Approximate is better than “exact” for
interval estimation of binomial proportions. The American Statistician,
52(2), 119-126.
}
\seealso{
\link{ci_p}.
}
\author{
Omar David Mercado Turizo, \email{omercado@unal.edu.co}
}
