% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_p_xxx.R
\name{ci_p_arcsine_ac}
\alias{ci_p_arcsine_ac}
\title{Arcsine Wald Confidence Interval with Continuity Correction Anscombe for Binomial Proportion}
\usage{
ci_p_arcsine_ac(x, n, conf.level = 0.95)
}
\arguments{
\item{x}{A number or a vector with the number of successes.}

\item{n}{A number or a vector with the number of trials.}

\item{conf.level}{Confidence level for the returned confidence interval. By default, it is 0.95.}
}
\value{
A vector with the lower and upper limits of the confidence interval.
}
\description{
This function calculates the Arcsine Wald confidence interval with continuity correction anscombe for a Binomial proportion.
It transforms the proportion to the arcsine scale, applies the Wald interval with continuity correction, and then back-transforms the result to the original scale.
The method is vectorized, allowing for evaluation of single values or vectors.
}
\details{
The Arcsine Wald confidence interval with continuity correction anscombe adjusts the classical Wald interval by transforming the proportion \eqn{\pi} to the arcsine scale.
The parameter \eqn{\pi} represents the true proportion of successes in a Binomial experiment, defined as:

\deqn{\pi = \frac{0.3875+x \pm 0.5}{n+0.75},}

where \eqn{x} is the number of successes and \eqn{n} is the number of trials.

On the arcsine scale, the transformed parameter is given by:

\deqn{\phi = \arcsin(\sqrt{\pi}).}

The standard error on the arcsine scale is constant:

\deqn{\text{se}(\phi) = \frac{1}{2\sqrt{n+0.5}},}

where \eqn{n} is the number of trials. The confidence interval on the arcsine scale is:

\deqn{\text{Lower}(\phi) = \max\left(0, \phi - z \cdot (\phi)\right),}
\deqn{\text{Upper}(\phi) = \min\left(\frac{\pi}{2}, \phi + z \cdot (\phi)\right),}

where \eqn{z} is the critical value from the standard normal distribution at the specified confidence level.

Back-transforming the limits to the original scale gives:

\deqn{\text{Lower}(\pi) = \sin^2(\text{Lower}(\phi)),}
\deqn{\text{Upper}(\pi) = \sin^2(\text{Upper}(\phi)).}

Special cases are handled explicitly:
- If \eqn{x = 0}, the lower limit is 0, and the upper limit is calculated as \eqn{(\alpha / 2)^{1/n}}.
- If \eqn{x = n}, the upper limit is 1, and the lower limit is calculated as \eqn{1 - (\alpha / 2)^{1/n}}.

These adjustments ensure that the confidence interval is valid and well-behaved, even at the boundaries of the parameter space.
}
\examples{
ci_p_arcsine_ac(x=15, n=50, conf.level=0.95)
ci_p_arcsine_ac(x= 0, n=50, conf.level=0.95)
ci_p_arcsine_ac(x=50, n=50, conf.level=0.95)

}
\references{
Missing reference.
}
\seealso{
\link{ci_p}.
}
\author{
David Esteban Cartagena Mejía, \email{dcartagena@unal.edu.co}
}
