% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_p_xxx.R
\name{ci_p_arcsine_anscombe}
\alias{ci_p_arcsine_anscombe}
\title{Arcsine Wald Confidence Interval with Continuity Correction Anscombe for
binomial proportion}
\usage{
ci_p_arcsine_anscombe(x, n, conf.level = 0.95)
}
\arguments{
\item{x}{A number or a vector with the number of successes.}

\item{n}{A number or a vector with the number of trials.}

\item{conf.level}{Confidence level for the returned confidence interval.
By default, it is 0.95.}
}
\value{
A vector with the lower and upper limits of the confidence interval.
}
\description{
This function calculates the confidence interval for a proportion.
It is vectorized, allowing users to evaluate it using either
single values or vectors.
}
\details{
The expression to obtain the confidence interval is given below:

\eqn{\sin^2(\arcsin(\sqrt{\tilde{p}}) \mp \frac{z_{\alpha/2}}{2\sqrt{n}})},

where \eqn{\tilde{p} = \frac{x+3/8}{n+3/4}}.
The value \eqn{z_{\alpha/2}} is the \eqn{1-\alpha/2} percentile of the
standard normal distribution (e.g., \eqn{z_{0.025}=1.96} for a 95\%
confidence interval).
}
\examples{
ci_p_arcsine_anscombe(x= 0, n=50, conf.level=0.95)
ci_p_arcsine_anscombe(x=15, n=50, conf.level=0.95)
ci_p_arcsine_anscombe(x=50, n=50, conf.level=0.95)

}
\references{
Anscombe, F. J. (1948). Transformations of Poisson, binomial and
negative-binomial data. Biometrika, 35, 246–254.
}
\seealso{
\link{ci_p}.
}
\author{
David Esteban Cartagena Mejía, \email{dcartagena@unal.edu.co}
}
