% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_p_coverage.R
\name{ci_p_coverage}
\alias{ci_p_coverage}
\title{Coverage for interval confidence p}
\usage{
ci_p_coverage(n, p, conf.level = 0.95, intervalType = "wald")
}
\arguments{
\item{n}{number of trials.}

\item{p}{true value for p.}

\item{conf.level}{nominal confidence level for the returned confidence interval. By default is 0.95.}

\item{intervalType}{type of confidence interval, possible choices are listed in \link{ci_p}.}
}
\value{
A dataframe with Method, n, p and true coverage.
}
\description{
This function calculates the true coverage for any interval confidence for p.
}
\details{
This function was inspired by the binomTestCoveragePlot() function from conf package and Park & Leemis (2019).
}
\examples{
ci_p_coverage(n=10, p=0.45, intervalType="wald")
ci_p_coverage(n=10, p=0.45, intervalType="wilson")
ci_p_coverage(n=10, p=c(0.10, 0.25, 0.40), intervalType="wilson")

}
\references{
Park, H., & Leemis, L. M. (2019). Ensemble confidence intervals for binomial proportions. Statistics in Medicine, 38(18), 3460-3475.
}
\seealso{
\link{ci_p}.
}
\author{
David Esteban Cartagena Mejía, \email{dcartagena@unal.edu.co}
}
