% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_p_coverage.R
\name{ci_p_coverage_plot}
\alias{ci_p_coverage_plot}
\title{Plot for confidence interval p - coverage vs p for fixed n}
\usage{
ci_p_coverage_plot(
  n,
  conf.level = 0.95,
  intervalType = "wald",
  plot = TRUE,
  seq_p = seq(from = 0.01, to = 0.99, length.out = 50),
  col = "deepskyblue2",
  linecolor = "tomato",
  ...
)
}
\arguments{
\item{n}{number of trials.}

\item{conf.level}{nominal confidence level for the returned confidence interval. By default is 0.95.}

\item{intervalType}{type of confidence interval, possible choices are listed in \link{ci_p}.}

\item{plot}{logical value to obtain the plot, TRUE by default.}

\item{seq_p}{sequence for the values of \eqn{n}. By default is \code{seq(from=0.01, to=0.99, length.out=50)}.}

\item{col}{color for the coverage curve.}

\item{linecolor}{color for the line representing the conf.level.}

\item{...}{further arguments and graphical parameters passed to plot function.}
}
\value{
A dataframe with Method, n, p and true coverage and the plot.
}
\description{
This function plots the coverage for any confidence interval for p.
}
\details{
This function was inspired by the binomTestCoveragePlot() function from conf package and Park & Leemis (2019).
}
\examples{
ci_p_coverage_plot(n=10,
                   intervalType="wald",
                   conf.level=0.95,
                   ylim=c(0.8, 1), las=1)

ci_p_coverage_plot(n=10,
                   intervalType="clopper_pearson",
                   conf.level=0.95,
                   ylim=c(0.9, 1), las=1)

ci_p_coverage_plot(n=10,
                   intervalType="wilson",
                   conf.level=0.95,
                   ylim=c(0.9, 1), las=1)

ci_p_coverage_plot(n=10,
                   intervalType="jeffreys",
                   conf.level=0.95,
                   ylim=c(0.9, 1), las=1)

ci_p_coverage_plot(n=10,
                   intervalType="agresti_coull",
                   conf.level=0.95,
                   ylim=c(0.9, 1), las=1)

ci_p_coverage_plot(n=10,
                   intervalType="arcsine",
                   conf.level=0.95,
                   ylim=c(0.9, 1), las=1)
}
\references{
Park, H., & Leemis, L. M. (2019). Ensemble confidence intervals for binomial proportions. Statistics in Medicine, 38(18), 3460-3475.
}
\seealso{
\link{ci_p}.
}
\author{
David Esteban Cartagena Mejía, \email{dcartagena@unal.edu.co}
}
