% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_p_coverage.R
\name{ci_p_coverage_plot2}
\alias{ci_p_coverage_plot2}
\title{Plot for confidence interval p - coverage vs n for fixed p}
\usage{
ci_p_coverage_plot2(
  p = 0.5,
  seq_n = seq(from = 10, to = 500, by = 10),
  conf.level = 0.95,
  intervalType = "wald",
  plot = TRUE,
  col = "deepskyblue2",
  linecolor = "tomato",
  ...
)
}
\arguments{
\item{p}{true proportion \eqn{p}.}

\item{seq_n}{sequence with the values of sample size \eqn{n}. By default is the sequence 10, 20, 30, ..., 480, 490, 500.}

\item{conf.level}{nominal confidence level for the returned confidence interval. By default is 0.95.}

\item{intervalType}{type of confidence interval, possible choices are listed in \link{ci_p}.}

\item{plot}{logical value to obtain the plot, TRUE by default.}

\item{col}{color for the coverage curve.}

\item{linecolor}{color for the line representing the conf.level.}

\item{...}{further arguments and graphical parameters passed to plot function.}
}
\value{
A dataframe with Method, n, p and true coverage and the plot.
}
\description{
This function plots the coverage for any confidence interval for p.
}
\details{
This function was inspired by the binomTestCoveragePlot() function from conf package and Park & Leemis (2019).
}
\examples{
ci_p_coverage_plot2(p=0.50,
                    conf.level=0.95,
                    intervalType="wald")

ci_p_coverage_plot2(p=0.75,
                    conf.level=0.95,
                    intervalType="wald")

ci_p_coverage_plot2(p=0.50,
                    conf.level=0.95,
                    intervalType="wilson")

ci_p_coverage_plot2(p=0.75,
                    conf.level=0.95,
                    intervalType="wilson",
                    col="orange", linecolor="purple")

}
\references{
Park, H., & Leemis, L. M. (2019). Ensemble confidence intervals for binomial proportions. Statistics in Medicine, 38(18), 3460-3475.
}
\seealso{
\link{ci_p}.
}
\author{
Laura Juliana Insignares Montes, \email{linsignares@unal.edu.co}
}
