% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_p_xxx.R
\name{ci_p_exact_clopper_pearson}
\alias{ci_p_exact_clopper_pearson}
\title{Exact confidence interval for Binomial proportion}
\usage{
ci_p_exact_clopper_pearson(x, n, conf.level = 0.95)
}
\arguments{
\item{x}{A number or a vector with the number of successes.}

\item{n}{A number or a vector with the number of trials.}

\item{conf.level}{Confidence level for the returned confidence interval. By default, it is 0.95.}
}
\value{
A vector with the lower and upper limits.
}
\description{
This function calculates the confidence interval for a proportion.
It is vectorized, allowing users to evaluate it using either
single values or vectors.
}
\details{
The exact confidence interval is based on the Clopper-Pearson method, which uses the F-distribution to calculate the limits. Special cases are handled when \eqn{x=0} or \eqn{x=n}.

The mathematical definitions are as follows:

- If \eqn{x=0}, the lower limit is \eqn{0}, and the upper limit is \eqn{1 - (\alpha / 2)^{1/n}}.

- If \eqn{x=n}, the lower limit is \eqn{(\alpha / 2)^{1/n}}, and the upper limit is \eqn{1}.

Otherwise, the limits are given by:

 \deqn{\text{Lower Limit}=\frac{1}{1+\frac{n-x+1}{x}F_{\alpha/2, \, 2(n-x+1), \, 2x}}}

 \deqn{\text{Upper Limit}=\frac{\frac{x+1}{n-x} F_{\alpha/2, \, 2(x+1), \, 2(n-x)}}{1+\frac{x+1}{n-x} F_{\alpha/2, \, 2(x+1), \, 2(n-x)}}}

where \eqn{F_{\omega, a, b}} is the \eqn{100\%(1-\omega)}
percentile of the F distribution with parameters
\eqn{a} and \eqn{b}.
}
\examples{
ci_p_exact_clopper_pearson(x= 0, n=50, conf.level=0.95)
ci_p_exact_clopper_pearson(x=15, n=50, conf.level=0.95)
ci_p_exact_clopper_pearson(x=50, n=50, conf.level=0.95)
ci_p_exact_clopper_pearson(x= 1, n=10, conf.level=0.95)

}
\references{
Blyth, C. R. (1986). Approximate binomial confidence limits. Journal of the American Statistical Association, 81(395), 843-855.
}
\seealso{
\link{ci_p}.
}
\author{
Omar David Mercado Turizo, \email{omercado@unal.edu.co}
}
