% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_p_xxx.R
\name{ci_p_hpd}
\alias{ci_p_hpd}
\title{Highest Posterior Density (HPD) interval for binomial proportion}
\usage{
ci_p_hpd(x, n, conf.level = 0.95, prior = "uniform")
}
\arguments{
\item{x}{A number or a vector with the number of successes.}

\item{n}{A number or a vector with the number of trials.}

\item{conf.level}{Confidence level for the returned confidence interval.
By default, it is 0.95.}

\item{prior}{The prior distribution to use. Options are
"uniform" (default) or "jeffreys".}
}
\value{
A vector with the lower and upper limits.
}
\description{
This function calculates the Highest Posterior Density (HPD) interval
for a binomial proportion using a Bayesian approach. It is vectorized,
allowing the evaluation of single values or vectors.
}
\details{
The HPD interval is a Bayesian credible interval for the Binomial
proportion \eqn{p}. The posterior distribution is calculated based
on the Beta prior:

- "uniform": \eqn{\text{Beta}(1, 1)}.

- "jeffreys": \eqn{\text{Beta}(0.5, 0.5)}.

The limits of the interval are computed using the quantiles
of the Beta posterior distribution:

- Lower limit: \eqn{\text{qbeta}((1 - \text{conf.level}) / 2, \alpha + x, \beta + n - x)}.

- Upper limit: \eqn{\text{qbeta}(1 - (1 - \text{conf.level}) / 2, \alpha + x, \beta + n - x)}.
}
\examples{
# Example with a single value
ci_p_hpd(x = 15, n = 50, conf.level = 0.95)
}
\references{
Missing reference.
}
\seealso{
\link{ci_p}.
}
\author{
Omar David Mercado Turizo, \email{omercado@unal.edu.co}
}
