% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_p_xxx.R
\name{ci_p_hpd_jeffreys}
\alias{ci_p_hpd_jeffreys}
\title{Highest Posterior Density (HPD) interval for binomial proportion using
Jeffreys prior}
\usage{
ci_p_hpd_jeffreys(x, n, conf.level = 0.95)
}
\arguments{
\item{x}{A number or a vector with the number of successes.}

\item{n}{A number or a vector with the number of trials.}

\item{conf.level}{Confidence level for the returned confidence interval.
By default, it is 0.95.}
}
\value{
A vector with the lower and upper limits of the HPD interval.
}
\description{
This function calculates the Highest Posterior Density (HPD) interval
for a Binomial proportion using Jeffreys prior. It is vectorized,
allowing the evaluation of single values or vectors.
}
\details{
The Jeffreys prior is a non-informative prior (\eqn{Beta(0.5, 0.5)})
based on the Fisher information. The posterior distribution is Beta:

\deqn{p | x \sim \text{Beta}(0.5 + x, 0.5 + n - x)}

The HPD interval is calculated using the posterior samples from the
Beta distribution:

\deqn{	\text{HPD Interval}=[L, U]}

where \eqn{L} and \eqn{U} are the bounds of the smallest interval
containing \eqn{1 - \alpha} posterior probability.
}
\examples{
# Example with a single value
ci_p_hpd_jeffreys(x=5, n=20, conf.level=0.95)

# Example with vectors
x_values <- c(5, 10, 15)
n_values <- c(20, 30, 40)
ci_p_hpd_jeffreys(x=x_values, n=n_values, conf.level=0.95)

}
\seealso{
\link{ci_p}.
}
\author{
Rusvelt Jose Meza San Martín, \email{rmezas@unal.edu.co}
}
