% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_p_xxx.R
\name{ci_p_jeffreys}
\alias{ci_p_jeffreys}
\title{Bayesian confidence interval for Binomial proportion using Jeffreys
prior (non-informative prior).}
\usage{
ci_p_jeffreys(x, n, conf.level = 0.95)
}
\arguments{
\item{x}{A number or a vector with the number of successes.}

\item{n}{A number or a vector with the number of trials.}

\item{conf.level}{Confidence level for the returned confidence interval.
By default, it is 0.95.}
}
\value{
A vector with the lower and upper limits.
}
\description{
This function calculates the confidence interval for a proportion.
It is vectorized, allowing users to evaluate it using either
single values or vectors.
}
\details{
The Jeffreys prior is a non-informative prior (\eqn{Beta(0.5, 0.5)})
based on the Fisher information. The posterior distribution is Beta:

\deqn{p | x \sim \text{Beta}(x+0.5, n-x+0.5)}

The limits of the Bayesian confidence interval are derived from the
quantiles of the posterior distribution:

\deqn{\text{Lower Limit}=B_{1-\alpha/2, x+0.5, n-x+0.5}}

\deqn{\text{Upper Limit}=B_{\alpha/2, x+0.5, n-x+0.5}}

where \eqn{B_{\omega, a, b}} is the \eqn{100\%(1-\omega)}
percentile of the Beta distribution with parameters
\eqn{a} and \eqn{b}.
}
\examples{
# Example with a single value
ci_p_jeffreys(x=5, n=20, conf.level=0.95)

# Example with vectors
x_values <- c(5, 10, 15)
n_values <- c(20, 30, 40)
ci_p_jeffreys(x=x_values, n=n_values, conf.level=0.95)

}
\seealso{
\link{ci_p}.
}
\author{
Rusvelt Jose Meza San Martin, \email{rmezas@unal.edu.co}
}
