% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_p_xxx.R
\name{ci_p_LRT}
\alias{ci_p_LRT}
\title{Likelihood Ratio confidence interval for binomial proportion}
\usage{
ci_p_LRT(x, n, conf.level = 0.95)
}
\arguments{
\item{x}{A number or a vector with the number of successes.}

\item{n}{A number or a vector with the number of trials.}

\item{conf.level}{Confidence level for the returned confidence interval.
By default, it is 0.95.}
}
\value{
A vector with the lower and upper limits.
}
\description{
This function calculates the Likelihood Ratio (LRT) confidence interval for
a binomial proportion.
It is vectorized, allowing the evaluation of single values or vectors.
}
\details{
This function computes the confidence interval for a binomial proportion
based on the Likelihood Ratio Test (LRT). The confidence interval is
defined as the set of values for \eqn{p} that satisfy the following
condition:

\deqn{-2 \log \left(\frac{L(p)}{L(\hat{p}_{ML})}\right) \leq \chi^2_{\gamma}(1),}

where \eqn{L(p)} is the likelihood function for the binomial model,
\eqn{\hat{p}_{ML} = x / n} is the maximum likelihood estimator
for \eqn{p}, and \eqn{\chi^2_{\gamma}(1)} is the \eqn{1-\gamma}
quantile of the chi-square distribution with 1
degree of freedom.

The confidence limits are calculated numerically using the \code{uniroot}
function in R. Special care is taken
to handle edge cases where \eqn{x = 0} or \eqn{x = n}, setting the
limits to 0 or 1, respectively.
}
\examples{
ci_p_LRT(x = 15, n = 50, conf.level = 0.95)

}
\references{
Somerville, M. C., & Brown, R. S. (2013). Exact likelihood ratio
and score confidence intervals for the binomial proportion.
Pharmaceutical statistics, 12(3), 120-128.
}
\seealso{
\link{ci_p}.
}
\author{
David Esteban Cartagena Mejía, \email{dcartagena@unal.edu.co}
}
