% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_p_xxx.R
\name{ci_p_score_cc}
\alias{ci_p_score_cc}
\title{Score confidence interval with continuity correction for Binomial proportion}
\usage{
ci_p_score_cc(x, n, conf.level = 0.95)
}
\arguments{
\item{x}{A number or a vector with the number of successes.}

\item{n}{A number or a vector with the number of trials.}

\item{conf.level}{Confidence level for the returned confidence interval.
By default, it is 0.95.}
}
\value{
A vector with the lower and upper limits.
}
\description{
This function calculates the score confidence interval with continuity
correction for a Binomial proportion. It is vectorized, allowing the
evaluation of single values or vectors.
}
\details{
The score confidence interval with continuity correction is an
adjusted interval for the Binomial proportion \eqn{p}.

The mathematical definitions are as follows:

Lower limit: \eqn{\frac{2np + z^2 - 1 - z \sqrt{z^2 - 2 - \frac{1}{n} + 4p(nq + 1)}}{2n + 2z^2}}.

Upper limit: \eqn{\frac{2np + z^2 + 1 + z \sqrt{z^2 + 2 - \frac{1}{n} + 4p(nq - 1)}}{2n + 2z^2}}.

Where \eqn{p = x / n} is the sample proportion, \eqn{q = 1 - p} its complement, and \eqn{z} is the critical value of the standard normal distribution.

The limits are truncated to the range \eqn{[0, 1]}.
}
\examples{
# Example with a single value
ci_p_score_cc(x = 15, n = 50, conf.level = 0.95)
}
\references{
Missing reference.
}
\seealso{
\link{ci_p}.
}
\author{
Omar David Mercado Turizo, \email{omercado@unal.edu.co}
}
