% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic_p_xxx.R
\name{ci_p_wald}
\alias{ci_p_wald}
\title{Wald confidence interval for Binomial proportion}
\usage{
ci_p_wald(x, n, conf.level = 0.95)
}
\arguments{
\item{x}{a number or a vector with the number of successes.}

\item{n}{a number or a vector with the number of trials.}

\item{conf.level}{confidence level for the returned confidence interval. By default is 0.95.}
}
\value{
A matrix with the lower and upper limits.
}
\description{
This function obtains the confidence interval for a proportion. It is vectorized, so the user can evaluate it using single values or a vector.
}
\examples{
ci_p_wald(x=15, n=50, conf.level=0.95)
}
\seealso{
\link{ci_p}.
}
\author{
Olga Bustos, \email{oabustos@unal.edu.co}
}
