% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_p_xxx.R
\name{ci_p_wald_bs}
\alias{ci_p_wald_bs}
\title{Wald with Blyth and Still approximation Binomial Score confidence
interval for binomial proportion}
\usage{
ci_p_wald_bs(x, n, conf.level = 0.95)
}
\arguments{
\item{x}{A number or a vector with the number of successes.}

\item{n}{A number or a vector with the number of trials.}

\item{conf.level}{Confidence level for the returned confidence interval.
By default, it is 0.95.}
}
\value{
A vector with the lower and upper limits of the confidence interval.
}
\description{
This function calculates the Wald Binomial Score confidence interval for
a binomial proportion.
It is vectorized, allowing the evaluation of single values or vectors.
}
\details{
The Wald Binomial Score confidence interval is an adjusted version of
the Wald interval,
designed to improve accuracy in small samples and near the boundaries
of the parameter space.

The estimated proportion is given by:

\deqn{\hat{p} = \frac{x}{n}},

and its complement is:

\deqn{\hat{q} = 1 - \hat{p}}.

The Wald Binomial Score interval incorporates adjustments for the sample
size and confidence level:

\deqn{\text{wn\_z} = \sqrt{n - z^2 - \frac{2z}{\sqrt{n}} - \frac{1}{n}}},

where \eqn{z} is the quantile of the standard normal distribution
corresponding to the confidence level.
The final confidence interval is given by:

\deqn{\text{Lower} = \hat{p} - \frac{z \sqrt{\hat{p}\hat{q} + \frac{1}{2n}}}{\text{wn\_z}},}
\deqn{\text{Upper} = \hat{p} + \frac{z \sqrt{\hat{p}\hat{q} + \frac{1}{2n}}}{\text{wn\_z}}.}

This interval is particularly useful when \eqn{n} is small or when
the proportion \eqn{\hat{p}} is close to 0 or 1.
}
\examples{
ci_p_wald_bs(x = 15, n = 50, conf.level = 0.95)

}
\references{
Blyth, C.R. and Still, H.A. (1983). Binomial confidence intervals,
Journal of the American Statistical Association, 78, 108–116.
}
\seealso{
\link{ci_p}.
}
\author{
David Esteban Cartagena Mejía, \email{dcartagena@unal.edu.co}
}
