% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_p_xxx.R
\name{ci_p_wald_cc}
\alias{ci_p_wald_cc}
\title{Wald Continuity-Corrected confidence interval for binomial proportion}
\usage{
ci_p_wald_cc(x, n, conf.level = 0.95)
}
\arguments{
\item{x}{A number or a vector with the number of successes.}

\item{n}{A number or a vector with the number of trials.}

\item{conf.level}{Confidence level for the returned confidence interval.
By default, it is 0.95.}
}
\value{
A vector with the lower and upper limits of the confidence interval.
}
\description{
This function calculates the Wald continuity-corrected confidence interval
for a binomial proportion.
It is vectorized, allowing the evaluation of single values or vectors.
}
\details{
The Wald continuity-corrected confidence interval adjusts the standard Wald interval for small sample sizes or when the proportion
\eqn{\hat{p}} is near 0 or 1. It incorporates a continuity correction to improve accuracy.

The estimated proportion is given by:

\deqn{\hat{p} = \frac{x}{n},}

and its complement is:

\deqn{\hat{q} = 1 - \hat{p}.}

The continuity-corrected interval incorporates the critical value \eqn{z} from the standard normal distribution:

\deqn{\text{Lower} = \hat{p} - z \sqrt{\frac{\hat{p}\hat{q}}{n}} - \frac{1}{2n},}
\deqn{\text{Upper} = \hat{p} + z \sqrt{\frac{\hat{p}\hat{q}}{n}} + \frac{1}{2n}.}


These adjustments ensure that the confidence interval is valid even
at the boundaries of the parameter space.
}
\examples{
ci_p_wald_cc(x = 15, n = 50, conf.level = 0.95)
ci_p_wald_cc(x = 0, n = 50, conf.level = 0.95)
ci_p_wald_cc(x = 50, n = 50, conf.level = 0.95)

}
\references{
Pires, Ana M., and Conceiçao Amado. "Interval estimators for a binomial
proportion: Comparison of twenty methods".
REVSTAT-Statistical Journal 6.2 (2008): 165-197.
}
\seealso{
\link{ci_p}.
}
\author{
David Esteban Cartagena Mejía, \email{dcartagena@unal.edu.co}
}
