% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_p_xxx.R
\name{ci_p_wald_recentered_cc}
\alias{ci_p_wald_recentered_cc}
\title{Recentered Wald Interval with Continuity Correction for binomial proportion}
\usage{
ci_p_wald_recentered_cc(x, n, conf.level = 0.95)
}
\arguments{
\item{x}{A number or a vector with the number of successes.}

\item{n}{A number or a vector with the number of trials.}

\item{conf.level}{Confidence level for the returned confidence interval.
By default, it is 0.95.}
}
\value{
A vector with the lower and upper limits of the confidence interval.
}
\description{
This function calculates the recentered Wald confidence interval with
continuity correction for a binomial proportion.
It adjusts the classical Wald interval by introducing a recentering term
and a continuity correction, improving accuracy for small sample sizes
and boundary cases.
The method is vectorized, allowing for evaluation of single
values or vectors.
}
\details{
The recentered Wald interval with continuity correction adjusts the
classical Wald interval by incorporating a recentering term and a
continuity correction to account for the discreteness of the
binomial distribution.

The critical value \eqn{z} is obtained from the standard normal
distribution for the specified confidence level:

\deqn{z = \Phi^{-1}(1 - \alpha / 2),}

where \eqn{\alpha = 1 - \text{conf.level}}.

The confidence limits are calculated as:

\deqn{\text{Lower} = \max\left(\frac{x + z^2 / 2}{n + z^2} - \left[z \sqrt{\frac{x}{n^2} \left(1 - \frac{x}{n}\right)} + \frac{1}{2n}\right], 0\right),}
\deqn{\text{Upper} = \min\left(\frac{x + z^2 / 2}{n + z^2} + \left[z \sqrt{\frac{x}{n^2} \left(1 - \frac{x}{n}\right)} + \frac{1}{2n}\right], 1\right).}

Special cases are handled explicitly:

- If \eqn{x = 0}, the lower limit is 0, and the upper limit is
calculated as \eqn{(\alpha / 2)^{1/n}}.

- If \eqn{x = n}, the upper limit is 1, and the lower limit is
calculated as \eqn{1 - (\alpha / 2)^{1/n}}.

These adjustments ensure that the confidence interval is valid and
well-behaved, even at the boundaries of the parameter space.
}
\examples{
ci_p_wald_recentered_cc(x =  0, n = 50, conf.level = 0.95)
ci_p_wald_recentered_cc(x = 25, n = 50, conf.level = 0.95)
ci_p_wald_recentered_cc(x = 50, n = 50, conf.level = 0.95)

}
\references{
Pires, Ana M., and Conceiçao Amado. "Interval estimators for a binomial
proportion: Comparison of twenty methods".
REVSTAT-Statistical Journal 6.2 (2008): 165-197.
}
\seealso{
\link{ci_p}.
}
\author{
David Esteban Cartagena Mejía, \email{dcartagena@unal.edu.co}
}
