% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_p_xxx.R
\name{ci_p_wald_t}
\alias{ci_p_wald_t}
\title{Wald-t Confidence Interval for Binomial Proportion}
\usage{
ci_p_wald_t(x, n, conf.level = 0.95)
}
\arguments{
\item{x}{A number or a vector with the number of successes.}

\item{n}{A number or a vector with the number of trials.}

\item{conf.level}{Confidence level for the returned confidence interval.
By default, it is 0.95.}
}
\value{
A vector with the lower and upper limits of the confidence interval.
}
\description{
This function calculates the Wald-t confidence interval for a Binomial
proportion using the method XIX proposed by Pan (2002).
It incorporates the use of a t-distribution with adjusted degrees of
freedom \eqn{\nu}, as specified in equation (2.8).
The function is vectorized, allowing for evaluation of single
values or vectors.
}
\details{
The Wald-t confidence interval is a modification of the classical
Wald interval that uses the t-distribution instead of the normal
distribution for greater accuracy in small samples. The estimated
proportion \eqn{\hat{p}} is adjusted as:

\deqn{\hat{p} = \frac{x + 2}{n + 4},}

which reduces bias in the interval estimation. The variance
 \eqn{V(\hat{p}, n)} is given by:

\deqn{V(\hat{p}, n) = \frac{\hat{p}(1 - \hat{p})}{n}.}

The degrees of freedom \eqn{\nu} are calculated using equation (2.8):

\deqn{\nu = \frac{2 V(\hat{p}, n)^2}{\Omega(\hat{p}, n)},}

where \eqn{\Omega(\hat{p}, n)} is defined as:

\deqn{\Omega(\hat{p}, n) = \frac{\hat{p} - \hat{p}^2}{n^3} + \frac{\hat{p} + (6n - 7)\hat{p}^2 + 4(n - 1)(n - 3)\hat{p}^3 - 2(n - 1)(2n - 3)\hat{p}^4}{n^5} - \frac{2(\hat{p} + (2n - 3)\hat{p}^2 - 2(n - 1)\hat{p}^3)}{n^4}.}

The confidence interval is then calculated as:

\deqn{\text{Lower} = \max\left(0, \hat{p} - t \cdot \sqrt{V(\hat{p}, n)}\right),}
\deqn{\text{Upper} = \min\left(1, \hat{p} + t \cdot \sqrt{V(\hat{p}, n)}\right),}

where \eqn{t} is the critical value from the t-distribution
with \eqn{\nu} degrees of freedom.
}
\examples{
ci_p_wald_t(x =  0, n = 50, conf.level = 0.95)
ci_p_wald_t(x = 15, n = 50, conf.level = 0.95)
ci_p_wald_t(x = 50, n = 50, conf.level = 0.95)

}
\references{
Pires, Ana M., and Conceiçao Amado. "Interval estimators for a binomial
proportion: Comparison of twenty methods".
REVSTAT-Statistical Journal 6.2 (2008): 165-197.
}
\seealso{
\link{ci_p}.
}
\author{
David Esteban Cartagena Mejía, \email{dcartagena@unal.edu.co}
}
