% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_p_xxx.R
\name{ci_p_wilson}
\alias{ci_p_wilson}
\title{Wilson confidence interval for Binomial proportion}
\usage{
ci_p_wilson(x, n, conf.level = 0.95)
}
\arguments{
\item{x}{a number or a vector with the number of successes.}

\item{n}{a number or a vector with the number of trials.}

\item{conf.level}{confidence level for the returned confidence interval.
By default is 0.95.}
}
\value{
A matrix with the lower and upper limits.
}
\description{
This function calculates the confidence interval for a proportion.
It is vectorized, allowing users to evaluate it using either
single values or vectors.
}
\details{
The expression to obtain the confidence interval is given below:

\eqn{\frac{\hat{p}+ \frac{z_{\alpha/2}^2}{2n}}{\widetilde{n}} \pm \frac{z_{\alpha/2}^2}{\widetilde{n}}  \sqrt{ (\hat{p}(1 - \hat{p}) + \frac{z_{\alpha/2}^2}{4n} )/n}},

where \eqn{\hat{p}=\frac{x}{n}} is the sample proportion, \eqn{\widetilde{n}=1 + \frac{ z_{\alpha/2}^2}{n}}, \eqn{x} the
number of observed successes in the sample with size \eqn{n}.
The value \eqn{z_{\alpha/2}} is the \eqn{1-\alpha/2} percentile of the
standard normal distribution (e.g., \eqn{z_{0.025}=1.96} for a 95\%
confidence interval).
}
\examples{
ci_p_wilson(x= 0, n=50, conf.level=0.95)
ci_p_wilson(x=15, n=50, conf.level=0.95)
ci_p_wilson(x=50, n=50, conf.level=0.95)

}
\references{
Wilson, E. B. (1927). Probable inference, the law of succession, and
statistical inference. Journal of the American Statistical Association,
22(158), 209-212.
}
\seealso{
\link{ci_p}.
}
\author{
Victor David Usuga, \email{vusuga@unal.edu.co}
}
