% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d.meantest.R
\name{d.meantest}
\alias{d.meantest}
\title{D-value for hypothesis test using vectors}
\usage{
d.meantest(x, y, alternative = "less", nrep = 1e+06)
}
\arguments{
\item{x}{vector sample data x.}

\item{y}{vector sample data y.}

\item{alternative}{a character string specifying the alternative hypothesis, must be one of \code{"less"} (default), \code{"greater"} or \code{"two.sided"}. You can specify just the initial letter.}

\item{nrep}{number of repetions to obtain the D-value, by default is 1000000}
}
\value{
A list with class \code{"htest"} containing the following 
 components:
\item{d.value}{the d-value for the test.}
\item{estimate}{the estimated mean or difference in means depending
      on whether it was a one-sample test or a two-sample test.}
\item{alternative}{a character string describing the alternative hypothesis.}
\item{method}{a character string indicating the type of test performed.}
}
\description{
This function shows the D-value of a hypothesis test for two samples using non-summarized data, that is, vectors.
}
\details{
In this case, it is important to explain that the sample size is assumed as n = 1 
because the D-value provides an individual analysis.
}
\examples{
d.meantest(x=c(2,3,4,5,6), y=c(3,4,5,6,7), alternative="two.sided")

}
