% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_meantest.R
\name{d_meantest}
\alias{d_meantest}
\title{D-value for hypothesis test using values}
\usage{
d_meantest(meanx, varx, meany, vary, alternative = "less",
  nrep = 1e+06)
}
\arguments{
\item{meanx}{sample mean for sample x.}

\item{varx}{sample variance for sample x.}

\item{meany}{sample mean for sample y.}

\item{vary}{sample variance for sample y.}

\item{alternative}{a character string specifying the alternative hypothesis, must be one of \code{"less"} (default), \code{"greater"} or \code{"two.sided"}. You can specify just the initial letter.}

\item{nrep}{number of repetions to obtain the D-value, by default is 1000000}
}
\value{
A list with class \code{"htest"} containing the following 
 components:
\item{d.value}{the d-value for the test.}
\item{estimate}{the estimated mean or difference in means depending
      on whether it was a one-sample test or a two-sample test.}
\item{alternative}{a character string describing the alternative hypothesis.}
\item{method}{a character string indicating the type of test performed.}
}
\description{
This function shows the D-value of a hypothesis test for two samples using summarized values, not the vectors.
}
\details{
In this case, it is important to explain that the sample size is assumed as n = 1 
because the D-value provides an individual analysis.
}
\examples{
# --- Examples with TWO-SAMPLES and equal variances ---
d_meantest(meanx=250, varx=20^2, meany=249, vary=20^2, alternative='greater')
 
}
