% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hola.R
\name{hola}
\alias{hola}
\title{Histograma rosado}
\usage{
hola(media, varianza, n)
}
\arguments{
\item{media}{es la media de la poblacion.}

\item{varianza}{la varianza de la poblacion.}

\item{n}{sample size.}
}
\description{
Sirve para crear un histograma de una muestra aleatoria de una normal.
}
\examples{
hola(media=170, varianza=10, n=59)
hola(media=-50, varianza=1, n=23)

}
\author{
Freddy Hernandez
}
