% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mi_densidad.R
\name{mi_densidad}
\alias{mi_densidad}
\title{Grafica una densidad}
\usage{
mi_densidad(n, media, varianza)
}
\arguments{
\item{n}{tamano de muestra.}

\item{media}{media de la poblacion.}

\item{varianza}{varianza de la poblacion}
}
\description{
sirve para graficar la densidad de una muestra de una normal.
}
\examples{
mi_densidad(n=10,media=0,varianza = 1)
mi_densidad(n=100,media = 10,varianza= 4)

}
