% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mi_hist.R
\name{mi_hist}
\alias{mi_hist}
\title{Crea un histograma rosado}
\usage{
mi_hist(n, media, varianza)
}
\arguments{
\item{n}{tamano de muestra.}

\item{media}{es la media de la poblacion.}

\item{varianza}{es la varianza de la poblacion.}
}
\description{
Sirve para construir un histograma de una muestra de una poblacion normal.
}
\examples{
mi_hist(n=10, media=15, varianza=3)
mi_hist(n=100, media=45, varianza=12)

}
