% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mi_qqnorm.R
\name{mi_qqnorm}
\alias{mi_qqnorm}
\title{Crea un grafico qqnorm}
\usage{
mi_qqnorm(n, media, varianza)
}
\arguments{
\item{n}{tamano de muestra}

\item{media}{media de la poblacion}

\item{varianza}{varianza de la poblacion}
}
\description{
Sirve para explorar la normalidad de una muestra mediante un gráfico cuantil cuantil
}
\examples{
mi_qqnorm(n = 30, media = 10, varianza = 4)
mi_qqnorm(n = 100, media = 25, varianza = 12)

}
