% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mu.test.R
\name{mu.test}
\alias{mu.test}
\title{Test for mean vector with covariance matrix known}
\usage{
mu.test(data, mu0, Sigma, alpha = 0.05)
}
\arguments{
\item{data}{a data frame.}

\item{mu0}{the reference vector, it could be an usual vector obtain with `c()` or a row/column matrix.}

\item{Sigma}{the covariance matrix.}

\item{alpha}{by default its value is 0.05.}
}
\value{
A list with the following components:
\item{x2}{the value of the statistic.}
\item{critic_value}{the critic value to make a decision.}
\item{pvalue}{the p-value for the test.}
}
\description{
This function performs the test for mean vector of a multivariate normal population with covariance matrix known.
}
\examples{
# Example 3.4.1 from Diaz & Morales (2016)
# H0: mu = (70, 170)
# H1: mu != (70, 170)
x1 <- c(69, 74, 68, 70, 72, 67, 66, 70, 76, 68,
        72, 79, 74, 67, 66, 71, 74, 75, 75, 76)
x2 <- c(153, 175, 155, 135, 172, 150, 115, 137, 200, 130,
        140, 265, 185, 112, 140, 150, 165, 185, 210, 220)
dt <- data.frame(x1, x2)

mu0 <- c(70, 170)
Sigma <- matrix(c(20, 100,
                  100, 1000), byrow=TRUE, ncol=2)
mu.test(data=dt, mu0=mu0, Sigma=Sigma, alpha=0.05)

}
