% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn_test.R
\name{mvn_test}
\alias{mvn_test}
\title{Modified Nel and Van der Merwe test}
\usage{
mvn_test(xbar1, Sigma1, n1, xbar2, Sigma2, n2)
}
\arguments{
\item{xbar1}{a vector with the sample mean from population 1.}

\item{Sigma1}{a matrix with sample variances and covariances from population 1.}

\item{n1}{sample size 1.}

\item{xbar2}{a vector with the sample mean from population 2.}

\item{Sigma2}{a matrix with sample variances and covariances from population 2.}

\item{n2}{sample size 2.}
}
\value{
A list with class \code{"htest"} containing the following components:
\item{statistic}{the value of the statistic.}
\item{parameter}{the degrees of freedom for the test.}
\item{p.value}{the p-value for the test.}
\item{estimate}{the estimated mean vectors.}
\item{method}{a character string indicating the type of test performed.}
}
\description{
This function implements the Modified Nel and Van der Merwe test the Behrens-Fisher problem \eqn{H0: \mu 1 = \mu 2} versus \eqn{H1: \mu1} not = \eqn{\mu2} without assuming anything about the covariance matrices.
}
\examples{
# Example 4.1 from Nel and Van de Merwe (1986) page 3729
# Test H0: mu1 = mu2 versus H1: mu1 != mu2
n1 <- 45
xb1 <- c(204.4, 556.6)
s1 <- matrix(c(13825.3, 23823.4, 23823.4, 73107.4), ncol=2)

n2 <- 55
xb2 <- c(130.0, 355.0)
s2 <- matrix(c(8632.0, 19616.7, 19616.7, 55964.5), ncol=2)

res1 <- mvn_test(xbar1=xb1, Sigma1=s1, n1=n1, xbar2=xb2, Sigma2=s2, n2=n2)
res1
plot(res1, from=5, to=10, shade.col='orange')

\dontrun{
# Example using simulated data -----
# Parameters for the simulation
n1 <- n2 <- 500
mu1 <- c(0, 0)
sigma1 <- matrix(c(1, 0.5, 0.5, 1), ncol=2)
mu2 <- c(0, 0)
sigma2 <- matrix(c(1, -0.5, -0.5, 1), ncol=2)

# Simulating the data
library(MASS)
dt1 <- mvrnorm(n1, mu=mu1, Sigma=sigma1)
dt2 <- mvrnorm(n2, mu=mu2, Sigma=sigma2)

Compositional::james(dt1, dt2, R=2)
mvn_test(xbar1=colMeans(dt1), Sigma1=var(dt1), n1=nrow(dt1),
         xbar2=colMeans(dt2), Sigma2=var(dt2), n2=nrow(dt2))
 }

}
\author{
Freddy Hernandez
}
