% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permu_test_SD.R
\name{permu_test_SD}
\alias{permu_test_SD}
\title{Permutation test using SD}
\usage{
permu_test_SD(x, x_sd, y, y_sd, alternative = "two.sided")
}
\arguments{
\item{x}{numeric vector of data values.}

\item{x_sd}{numeric vector with standard deviations.}

\item{y}{numeric vector of data values.}

\item{y_sd}{numeric vector with standard deviations.}

\item{alternative}{a character string specifying the alternative
hypothesis, must be one of \code{"two.sided"} (default),
\code{"greater"} or \code{"less"}. You can specify just the initial letter.}
}
\value{
A list with class \code{"htest"} containing the following
 components:
\item{statistic}{the value of the statistic.}
\item{p.value}{the p-value for the test.}
\item{alternative}{a character string describing the alternative hypothesis.}
\item{method}{a character string indicating the type of test performed.}
}
\description{
Permutation test using SD
}
\details{
This function is based in the Python function proposed
by Ermolinskiy (2024).
}
\examples{
# Example 0
x <- c(5, 6)
x_sd <- c(1, 2)
y <- c(7, 9)
y_sd <- c(1, 2)

res <- permu_test_SD(x, x_sd, y, y_sd, alternative="less")
res

# Example 1 of the paper
x <- c(20, 33, 20, 24)
x_sd <- c(2, 2, 2, 2)
y <- c(21, 34, 22, 25)
y_sd <- c(2, 2, 2, 2)

res <- permu_test_SD(x, x_sd, y, y_sd, alternative="less")
res

}
\references{
Ermolinskiy, P. (2024). An Extension of the Permutation Test for
Dependent Data in Samples with a Small Number of Subjects.
https://hal.science/hal-04558513/
}
\author{
Freddy Hernandez, \email{fhernanb@unal.edu.co}

This function performs test for dependent data in samples
with a small number of subjects.
}
