% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.htest.R
\name{plot.htest}
\alias{plot.htest}
\title{To plot p-value area}
\usage{
\method{plot}{htest}(x, col = "red", shade.col = "red", cex = 0.8,
  from = NULL, to = NULL, ...)
}
\arguments{
\item{x}{Object of class \dQuote{htest}.}

\item{col}{color for the observed statistic.}

\item{shade.col}{color for the shaded area.}

\item{cex}{a numerical value giving the amount by which plotting the p-value.}

\item{from}{the minimum value of the X-axis.}

\item{to}{the maximum value of the X-axis.}

\item{...}{Other plotting parameters to affect the plot.}
}
\value{
None. Function produces a plot.
}
\description{
This is a generic function to depict the p-value area for a object with class \code{htest}.
}
\examples{

# Example
# H0: mu = 170
# Ha: mu != 170 with sigma2=25
x <- rnorm(n=80, mean=171, sd=5)
res <- z.test(x=x, mu=170, sigma2=25, alternative='two.sided')
res
plot(res, col='blue', shade.col='tomato')

}
\keyword{plot.htest}
