% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suma.R
\name{suma}
\alias{suma}
\title{suma}
\usage{
suma(x, y)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Esta funcion sirve para obtener la suma de dos numeros reales.
}
\examples{
suma(1, 1)
suma(10, 1)
}
