% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_mean_vector_test.R
\name{two_mean_vector_test}
\alias{two_mean_vector_test}
\title{Tests for Equality of Two Normal Mean Vectors}
\usage{
two_mean_vector_test(
  xbar1,
  s1,
  n1,
  xbar2,
  s2,
  n2,
  delta0 = NULL,
  method = "T2",
  alpha = 0.05
)
}
\arguments{
\item{xbar1}{a vector with the sample mean from population 1.}

\item{s1}{a matrix with sample variances and covariances from population 1.}

\item{n1}{sample size 1.}

\item{xbar2}{a vector with the sample mean from population 2.}

\item{s2}{a matrix with sample variances and covariances from population 2.}

\item{n2}{sample size 2.}

\item{delta0}{a number indicating the true value of the difference in means.}

\item{method}{a character string specifying the method, \code{"T2"} (default), \code{"james"} (James' first order test), please see the details section for other methods.}

\item{alpha}{the significance level only for method \code{"james"}, by default its value is 0.05.}
}
\value{
A list with class \code{"htest"} containing the following components:
\item{statistic}{the value of the statistic.}
\item{parameter}{the degrees of freedom for the test.}
\item{p.value}{the p-value for the test.}
\item{estimate}{the estimated mean vectors.}
\item{method}{a character string indicating the type of test performed.}
}
\description{
The function implements the test for \eqn{H_0: \mu_1 = \mu_2} versus \eqn{H_1: \mu_1} not = \eqn{\mu_2} when two random samples are obtained from two p-variate normal populations \eqn{Np(\mu_1, \Sigma_1)} and \eqn{Np(\mu_2, \Sigma_2)} respectively. By default, the function performs the Hotelling test for two normal mean vectors assumming equality in the covariance matrices. Other tests for the multivariate Behrens-Fisher problem are also implemented, see the argument \code{method} below.
}
\details{
the \code{"method"} must be one of \code{"T2"} (default), \code{"james"} (James' first order test), \code{"yao"} (Yao's test), \code{"johansen"} (Johansen's test), \code{"nvm"} (Nel and Van der Merwe test), \code{"mnvm"} (modified Nel and Van der Merwe test), \code{"gamage"} (Gamage's test), \code{"yy"} (Yanagihara and Yuan test), \code{"byy"} (Bartlett Correction test), \code{"ks1"} (Kawasaki and Seo). For James test the critic value is reported, we reject H0 if T2 > critic_value.
}
\examples{
# Example 5.4.2 from Rencher & Christensen (2012) page 137,
# using Hotelling's test
n1 <- 32
xbar1 <- c(15.97, 15.91, 27.19, 22.75)
s1 <- matrix(c(5.192, 4.545, 6.522, 5.25,
               4.545, 13.18, 6.76, 6.266,
               6.522, 6.76, 28.67, 14.47,
               5.25, 6.266, 14.47, 16.65), ncol = 4)

n2 <- 32
xbar2 <- c(12.34, 13.91, 16.66, 21.94)
s2 <- matrix(c(9.136, 7.549, 4.864, 4.151,
               7.549, 18.6, 10.22, 5.446,
               4.864, 10.22, 30.04, 13.49,
               4.151, 5.446, 13.49, 28), ncol = 4)

res1 <- two_mean_vector_test(xbar1 = xbar1, s1 = s1, n1 = n1,
                             xbar2 = xbar2, s2 = s2, n2 = n2,
                             method = "T2")
res1
plot(res1, from=21, to=25, shade.col='tomato')


}
\author{
Freddy Hernandez, Jean Paul Piedrahita, Valentina Garcia.
}
