% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var.test.R
\name{var.test}
\alias{var.test}
\title{Variance test using vectors}
\usage{
var.test(
  x,
  y = NULL,
  alternative = "two.sided",
  null.value = 1,
  conf.level = 0.95
)
}
\arguments{
\item{x}{a (non-empty) numeric vector of data values.}

\item{y}{an optional (non-empty) numeric vector of data values.}

\item{alternative}{a character string specifying the alternative
hypothesis, must be one of \code{two.sided} (default),
\code{greater} or \code{less}. You can specify just the initial letter.}

\item{null.value}{the hypothesized number (variance or ratio of the variances) in the null hypothesis.}

\item{conf.level}{confidence level of the interval, by default its value is 0.95.}
}
\value{
A list with class \code{htest} containing the following
 components:
\item{statistic}{the value of the statistic.}
\item{p.value}{the p-value for the test.}
\item{conf.int}{a confidence interval for the variance.}
\item{estimate}{the sample variance (or ratio of the sample variances)}
\item{null.value}{the specified hypothesized value for alternative hypothesis.}
\item{alternative}{a character string describing the alternative hypothesis.}
\item{method}{a character string indicating the type of test performed.}
\item{data.name}{a character string giving the name of the data.}
}
\description{
This function performs the test for a single variance or two variances given the vectors. This function is a generalization of \code{var.test} function from \code{stats} package.
}
\examples{
# One sample -----

# Interval confidence
duration <- c(1470, 1510, 1690, 1740, 1900, 2000, 2030,
              2010, 2190, 2200, 2290, 2380, 2390, 2480,
              2500, 2580, 2700)
var.test(x=duration, conf.level=0.95)

# Hypothesis testing
#   H0: sigma2 = 100
#   H1: sigma2 > 100
weight <- c(775, 780, 781, 795, 803, 810, 823)
var.test(x=weight, alternative='greater', null.value=100)

# Using the plot function
res1 <- var.test(x=weight, alternative='greater', null.value=100)
plot(res1)

# Two samples -----

# Hypothesis testing
#   H0: sigma1/sigma2 = 1
#   H1: sigma1/sigma2 != 1
x1 <- rnorm(50, mean = 0, sd = 2)
x2 <- rnorm(30, mean = 1, sd = 1)
res2 <- var.test(x1, x2)
res2
plot(res2)

}
