% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_test.R
\name{var_test}
\alias{var_test}
\title{Variance test using values}
\usage{
var_test(
  varx,
  nx,
  vary = NULL,
  ny = NULL,
  alternative = "two.sided",
  null.value = 1,
  conf.level = 0.95
)
}
\arguments{
\item{varx}{sample variance for sample x.}

\item{nx}{sample size for sample x.}

\item{vary}{sample variance for sample y.}

\item{ny}{sample size for sample y.}

\item{alternative}{a character string specifying the alternative
hypothesis, must be one of \code{two.sided} (default),
\code{greater} or \code{less}. You can specify just the initial letter.}

\item{null.value}{the hypothesized number (variance or ratio of the variances) in the null hypothesis.}

\item{conf.level}{confidence level of the interval, by default its value is 0.95.}
}
\value{
A list with class \code{htest} containing the following
 components:
\item{statistic}{the value of the statistic.}
\item{p.value}{the p-value for the test.}
\item{conf.int}{a confidence interval for the variance.}
\item{estimate}{the sample variance (or ratio of the sample variances)}
\item{null.value}{the specified hypothesized value for alternative hypothesis.}
\item{alternative}{a character string describing the alternative hypothesis.}
\item{method}{a character string indicating the type of test performed.}
}
\description{
This function performs the test for a single variance or two variances using values, not the vectors.
}
\examples{
# Examples with ONE sample

# Example 7.7.1 from Wayne (2013), http://tinyurl.com/y6z49hrw
var_test(varx=670.81, nx=16, null.value=600, alternative='two.sided')

# Exercise 7.7.5 from Wayne (2013), http://tinyurl.com/y6z49hrw
var_test(varx=30, nx=25, null.value=25, alternative='greater')

# Using the plot to illustrate Hypothesis Test
mytest1 <- var_test(varx=30, nx=25, null.value=25, alternative='greater')
mytest1
plot(mytest1)

# Examples with TWO samples

# Example 7.8 from Montgomery (1996)
var_test(varx=5.1^2, nx=12, vary=4.7^2, ny=15, conf.level=0.90)

# Example 8.17 from Montgomery (1996)
mytest2 <- var_test(varx=3.84, nx=20, vary=4.54, ny=20)
mytest2
plot(mytest2)

}
