% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_test.R
\name{z_test}
\alias{z_test}
\title{Mean test using values}
\usage{
z_test(
  meanx,
  nx,
  sigma2 = NULL,
  alternative = "two.sided",
  mu = 0,
  conf.level = 0.95
)
}
\arguments{
\item{meanx}{sample mean for sample x.}

\item{nx}{sample size for sample x.}

\item{sigma2}{population variance which is known.}

\item{alternative}{a character string specifying the alternative 
hypothesis, must be one of \code{two.sided} (default), 
\code{greater} or \code{less}. You can specify just the initial letter.}

\item{mu}{the hypothesized number in the null hypothesis.}

\item{conf.level}{confidence level of the interval, by default its value is 0.95.}
}
\value{
A list with class \code{htest} containing the following 
 components:
\item{statistic}{the value of the statistic.}
\item{p.value}{the p-value for the test.}
\item{conf.int}{a confidence interval for the variance.}
\item{estimate}{the estimated mean.}
\item{null.value}{the specified hypothesized value for alternative hypothesis.}
\item{alternative}{a character string describing the alternative hypothesis.}
\item{method}{a character string indicating the type of test performed.}
\item{data.name}{a character string giving the name of the data.}
}
\description{
This function performs the mean test using summarized data (values).
}
\examples{
# Example 13.1 from Freund et al. (2000)
res1 <- z_test(meanx=8.091, nx=25, mu=8, sigma2=0.16^2, alternative='two.sided')
res1
plot(res1)

# Example 13.2 from Freund et al. (2000)
res2 <- z_test(meanx=21819, nx=100, mu=22000, sigma2=1295^2, alternative='less')
res2
plot(res2)

}
