% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_p_xxx.R
\name{ci_p_arcsine}
\alias{ci_p_arcsine}
\title{ArcoSeno confidence interval for Binomial proportion}
\usage{
ci_p_arcsine(x, n, conf.level = 0.95)
}
\arguments{
\item{x}{a number or a vector with the number of successes.}

\item{n}{a number or a vector with the number of trials.}

\item{conf.level}{confidence level for the returned confidence interval. By default is 0.95.}
}
\value{
A matrix with the lower and upper limits.
}
\description{
This function calculates the confidence interval for a proportion.
It is vectorized, allowing users to evaluate it using either
single values or vectors.
}
\details{
The expression to obtain the confidence interval is given below:

\eqn{\sin^2(\hat\phi_{ML} \pm z_{\alpha/2} \times se(\hat\phi_{ML}))},

where the maximum likelihood estimator for \eqn{\phi} is
\eqn{\hat\phi_{ML}=\arcsin(\sqrt{\hat \pi_{ML}})}, and its standard error is
\eqn{se(\hat{\phi}_{ML})=\frac{1}{\sqrt{4n}}}. The
value \eqn{z_{\alpha/2}} is the \eqn{1-\alpha/2} percentile of the
standard normal distribution (e.g., \eqn{z_{0.025}=1.96} for a 95\%
confidence interval).
}
\examples{
ci_p_arcsine(x=15, n=50, conf.level=0.95)

}
\references{
Missing reference.
}
\seealso{
\link{ci_p}.
}
\author{
Victor David Usuga Duque, \email{vusuga@unal.edu.co}
}
