% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_p_xxx.R
\name{ci_p_clopper_pearson}
\alias{ci_p_clopper_pearson}
\title{Clopper-Pearson confidence interval for Binomial proportion}
\usage{
ci_p_clopper_pearson(x, n, conf.level = 0.95)
}
\arguments{
\item{x}{A number or a vector with the number of successes.}

\item{n}{A number or a vector with the number of trials.}

\item{conf.level}{Confidence level for the returned confidence interval. By default, it is 0.95.}
}
\value{
A vector with the lower and upper limits.
}
\description{
This function calculates the confidence interval for a proportion.
It is vectorized, allowing users to evaluate it using either
single values or vectors.
}
\details{
The Clopper-Pearson interval is an exact confidence interval for the Binomial proportion \eqn{p}.
The limits of the interval are derived based on the Beta distribution. For the special cases where \eqn{x=0} or \eqn{x=n}, the limits are calculated directly.

The mathematical definitions are as follows:

- If \eqn{x=0}, the lower limit is \eqn{0}, and the upper limit is \eqn{1 - (\alpha / 2)^{1/n}}.

- If \eqn{x=n}, the lower limit is \eqn{(\alpha / 2)^{1/n}}, and the upper limit is \eqn{1}.

Otherwise, the limits are given by:

\deqn{\text{Lower Limit}=B_{1-\alpha/2, x, n-x+1}}

\deqn{\text{Upper Limit}=B_{\alpha/2, x+1, n-x}}

where \eqn{B_{\omega, a, b}} is the \eqn{100\%(1-\omega)}
percentile of the Beta distribution with parameters
\eqn{a} and \eqn{b}.
}
\examples{
ci_p_clopper_pearson(x=15, n=50, conf.level=0.95)

}
\references{
Clopper, C. J., & Pearson, E. S. (1934). The use of confidence or fiducial limits illustrated in the case of the binomial. Biometrika, 26(4), 404-413.
}
\seealso{
\link{ci_p}.
}
\author{
Omar David Mercado Turizo, \email{omercado@unal.edu.co}
}
