% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_p_xxx.R
\name{ci_p_wald}
\alias{ci_p_wald}
\title{Wald confidence interval for Binomial proportion}
\usage{
ci_p_wald(x, n, conf.level = 0.95)
}
\arguments{
\item{x}{a number or a vector with the number of successes.}

\item{n}{a number or a vector with the number of trials.}

\item{conf.level}{confidence level for the returned confidence interval. By default is 0.95.}
}
\value{
A matrix with the lower and upper limits.
}
\description{
This function calculates the confidence interval for a proportion.
It is vectorized, allowing users to evaluate it using either
single values or vectors.
}
\details{
The expression to obtain the confidence interval is given below:

\eqn{\hat{p} - z_{\alpha/2} \sqrt{\frac{\hat{p}(1 - \hat{p})}{n}} \leq p \leq \hat{p} + z_{\alpha/2} \sqrt{\frac{\hat{p}(1 - \hat{p})}{n}}},

where \eqn{\hat{p}=\frac{x}{n}} is the sample proportion, \eqn{x} the
number of observed successes in the sample with size \eqn{n}. The
value \eqn{z_{\alpha/2}} is the \eqn{1-\alpha/2} percentile of the
standard normal distribution (e.g., \eqn{z_{0.025}=1.96} for a 95\%
confidence interval).
}
\examples{
ci_p_wald(x=15, n=50, conf.level=0.95)

}
\references{
Wald, A. (1949). Statistical decision functions. The Annals of Mathematical Statistics, 165-205.
}
\seealso{
\link{ci_p}.
}
\author{
Olga Bustos, \email{oabustos@unal.edu.co}
}
