% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_p_xxx.R
\name{ci_p_wald_bs}
\alias{ci_p_wald_bs}
\title{Wald with Blyth and Still approximation Binomial Score confidence
interval for binomial proportion}
\usage{
ci_p_wald_bs(x, n, conf.level = 0.95)
}
\arguments{
\item{x}{A number or a vector with the number of successes.}

\item{n}{A number or a vector with the number of trials.}

\item{conf.level}{Confidence level for the returned confidence interval.
By default, it is 0.95.}
}
\value{
A vector with the lower and upper limits of the confidence interval.
}
\description{
This function calculates the Wald Binomial Score confidence interval for
a binomial proportion.
It is vectorized, allowing the evaluation of single values or vectors.
}
\details{
The Wald Binomial Score confidence interval is an adjusted version of
the Wald interval,
designed to improve accuracy in small samples and near the boundaries
of the parameter space.

The bounds for the interval confidence are:

\eqn{\hat{p} \mp \left[ \frac{Z_{\alpha/2} \sqrt{\hat{p} (1-\hat{p})}}{\sqrt{n - Z_{\alpha/2}^2 - \frac{2Z_{\alpha/2}}{\sqrt{n}} - \frac{1}{n}}} + \frac{1}{2n} \right]}

This interval is particularly useful when \eqn{n} is small or when
the proportion \eqn{\hat{p}} is close to 0 or 1.
}
\examples{
ci_p_wald_bs(x = 15, n = 50, conf.level = 0.95)

}
\references{
Blyth, C.R. and Still, H.A. (1983). Binomial confidence intervals,
Journal of the American Statistical Association, 78, 108–116.
}
\seealso{
\link{ci_p}.
}
\author{
David Esteban Cartagena Mejía, \email{dcartagena@unal.edu.co}
}
