% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_mean_vector_test.R
\name{one_mean_vector_test}
\alias{one_mean_vector_test}
\title{Test for \eqn{\mu} in a \eqn{Np(\mu, \Sigma)}}
\usage{
one_mean_vector_test(mu0, xbar, n, S = NULL, Sigma = NULL)
}
\arguments{
\item{mu0}{a vector indicating the hypothesized value of the mean.}

\item{xbar}{a vector with the sample mean.}

\item{n}{sample size.}

\item{S}{a matrix with sample variances and covariances.}

\item{Sigma}{the matrix \eqn{\Sigma} if known.}
}
\value{
A list with class \code{"htest"}.
}
\description{
This function can be used to test \eqn{H0: \mu = \mu0} versus \eqn{H1: \mu} not = \eqn{\mu0}.
}
\details{
The user must provide only one matrix, S to perform the T2 test or Sigma to perform the X2 test.
}
\examples{
# Example 5.2.2 from Rencher & Christensen (2012) page 127
# Test H0: mu = (70, 170) versus H1: mu != (70, 170)
# with known Sigma

Sigma <- matrix(c(20, 100, 100, 1000), ncol=2, nrow=2)

res1 <- one_mean_vector_test(mu0=c(70, 170), xbar=c(71.45, 164.7),
                             n=20, Sigma=Sigma)
res1
plot(res1, from=4, to=10, shade.col='dodgerblue2')

# Example 5.2 from Johnson and Wichern (2012) page 214
# Test H0: mu = (4, 50, 10) versus H1: mu != (4, 50, 10)
# with unknown Sigma

S <- matrix(c(2.879, 10.010, -1.810,
              10.010, 199.788, -5.640,
              -1.810, -5.640, 3.628), ncol=3, nrow=3)

res2 <- one_mean_vector_test(mu0=c(4, 50, 10),
                             xbar=c(4.640, 45.400, 9.965),
                             n=20, S=S)
res2
plot(res2, from=0, to=5, shade.col='aquamarine3')

\dontrun{
library(rrcov)
data(delivery)
delivery.x <- delivery[, 1:2]

# Using T2.test from rrcov package
T2.test(delivery.x)

one_mean_vector_test(mu0=c(0, 0), xbar=colMeans(delivery.x),
                     n=25, S=var(delivery.x))
}

}
\seealso{
[one_covar_matrix_test()] for test \eqn{\Sigma} in a \eqn{Np(\mu, \Sigma)}.
}
\author{
Freddy Hernandez
}
